Fix types

+++ tyuiu.cources.programming.interfaces/ISprint3Task0.cs
@@ -6,19 +6,19 @@ using System.Threading.Tasks;
 
 namespace tyuiu.cources.programming.interfaces.Sprint3
 {
-    public interface ISprint3Task0V0 { int GetSumSeries(int value); }
-    public interface ISprint3Task0V1 { int GetSumSeries(int value); }
+    public interface ISprint3Task0V0 { double GetSumSeries(double value); }
+    public interface ISprint3Task0V1 { double GetSumSeries(double value); }
     public interface ISprint3Task0V2 { double GetSumSeries(); }
     public interface ISprint3Task0V3 { double GetSumSeries(); }
     public interface ISprint3Task0V4 { double GetSumSeries(); }
     public interface ISprint3Task0V5 { double GetSumSeries(); }
-    public interface ISprint3Task0V6 { double GetSumSeries(int value); }
+    public interface ISprint3Task0V6 { double GetSumSeries(double value); }
     public interface ISprint3Task0V7 { double GetSumSeries(double value); }    
     public interface ISprint3Task0V8 { double GetSumSeries(double value); }
     public interface ISprint3Task0V9 { double GetSumSeries(double value); }
-    public interface ISprint3Task0V10 { double GetSumSeries(int value); }
-    public interface ISprint3Task0V11 { double GetSumSeries(int value); }
-    public interface ISprint3Task0V12 { double GetSumSeries(int value); }
+    public interface ISprint3Task0V10 { double GetSumSeries(double value); }
+    public interface ISprint3Task0V11 { double GetSumSeries(double value); }
+    public interface ISprint3Task0V12 { double GetSumSeries(double value); }
     public interface ISprint3Task0V13 { double GetSumSeries(double value); }
     public interface ISprint3Task0V14 { double GetSumSeries(double value); }
     public interface ISprint3Task0V15 { double GetSumSeries(double value); }
@@ -30,10 +30,10 @@ namespace tyuiu.cources.programming.interfaces.Sprint3
     public interface ISprint3Task0V21 { double GetSumSeries(double value); }
     public interface ISprint3Task0V22 { double GetSumSeries(double value); }
     public interface ISprint3Task0V23 { double GetSumSeries(double value); }
-    public interface ISprint3Task0V24 { double GetSumSeries(int value); }
-    public interface ISprint3Task0V25 { double GetSumSeries(int value); }
-    public interface ISprint3Task0V26 { double GetSumSeries(int value); }
-    public interface ISprint3Task0V27 { double GetSumSeries(int value); }
+    public interface ISprint3Task0V24 { double GetSumSeries(double value); }
+    public interface ISprint3Task0V25 { double GetSumSeries(double value); }
+    public interface ISprint3Task0V26 { double GetSumSeries(double value); }
+    public interface ISprint3Task0V27 { double GetSumSeries(double value); }
     public interface ISprint3Task0V28 { double GetSumSeries(double value); }
     public interface ISprint3Task0V29 { double GetSumSeries(double value); }
     public interface ISprint3Task0V30 { double GetSumSeries(double value); }
